requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Pillow==10.4.0
gunicorn==22.0.0
python-dotenv==1.0.1
